+++ examples/racecar/common/eve_assetload_array.c
@@ -95,7 +95,7 @@ int eve_loadpatch_impl(void)
     EVE_LIB_BeginCoProList();
     EVE_CMD_LOADPATCH(0);
     do {
-        chunk = (size_t)(patch_asset.Array_Size - total);
+        chunk = (uint32_t)(patch_asset.Array_Size - total);
         if (chunk > 256) 
         {
             chunk = 256;
@@ -127,7 +127,7 @@ void eve_asset_load(EVE_ASSET_PROPS *asset, uint32_t loadimage)
         EVE_LIB_BeginCoProList();
         EVE_CMD_LOADIMAGE(asset->RAM_G_Start, 0);
         do {
-            chunk = (size_t)(asset->Array_Size - total);
+            chunk = (uint32_t)(asset->Array_Size - total);
             if (chunk > 256) 
             {
                 chunk = 256;
@@ -182,7 +182,7 @@ void eve_asset_load(EVE_ASSET_PROPS *asset, uint32_t loadimage)
         EVE_LIB_BeginCoProList();
         EVE_CMD_MEMWRITE(asset->RAM_G_Start, asset->Array_Size);
         do {
-            chunk = (size_t)(asset->Array_Size - total);
+            chunk = (uint32_t)(asset->Array_Size - total);
             if (chunk > 256) 
             {
                 chunk = 256;
